Removed duplicated code in pges
